==> an-l16/0_prep_download_unzip.ipynb <==
!rm -rf downloads

!mkdir downloads

!wget -P downloads https://data.assemblee-nationale.fr/static/openData/repository/16/amo/deputes_senateurs_ministres_legislature/AMO20_dep_sen_min_tous_mandats_et_organes.json.zip

!wget -P downloads https://data.assemblee-nationale.fr/static/openData/repository/16/loi/scrutins/Scrutins.json.zip

!wget -P downloads https://data.assemblee-nationale.fr/static/openData/repository/16/loi/amendements_div_legis/Amendements.json.zip

!wget -P downloads https://data.assemblee-nationale.fr/static/openData/repository/16/loi/dossiers_legislatifs/Dossiers_Legislatifs.json.zip

!wget -P downloads https://data.assemblee-nationale.fr/static/openData/repository/16/vp/reunions/Agenda.json.zip

!wget -P downloads https://data.assemblee-nationale.fr/static/openData/repository/16/vp/syceronbrut/syseron.xml.zip

!rm -rf data

!mkdir data

!unzip downloads/AMO20_dep_sen_min_tous_mandats_et_organes.json.zip -d data

!unzip downloads/Scrutins.json.zip -d data

!unzip downloads/Amendements.json.zip -d data

!unzip downloads/Dossiers_Legislatifs.json.zip -d data

!mkdir -p data/reunions

!unzip downloads/Agenda.json.zip -d data/reunions

!mkdir -p data/debats

!unzip downloads/syseron.xml.zip -d data/debats
