==> an-l16/1_prep_amendements.ipynb <==
import pandas as pd
import numpy as np
import json
import pandas as pd
from glob import glob
import tqdm

import dagstermill

def get_amendements(filename):
    return json.load(open(filename))['amendement']

amendements_json = [ get_amendements(f) for f in glob('data/json/**/**/AMAN*.json') ]

len(amendements_json)

def parse_amendement(a):
    return {
        'id': a['uid'],
        'numero': a['identification']['numeroLong'],
        'texte': a['texteLegislatifRef'],
        'signataire': a['signataires']['auteur']['acteurRef'],
        'organe': a['signataires']['auteur']['groupePolitiqueRef'],
        'dispositif' : a['corps']['contenuAuteur']['dispositif'] if 'contenuAuteur' in a['corps'] and 'dispositif' in a['corps']['contenuAuteur'] else np.nan,
        'seance': a['seanceDiscussionRef'] if '@xsi:nil' not in a['seanceDiscussionRef'] else np.nan,
        'sort': a['cycleDeVie']['sort'] if '@xsi:nil' not in a['cycleDeVie']['sort'] else np.nan,
        'discussion_commune': a['discussionCommune'] if '@xsi:nil' not in a['discussionCommune'] else np.nan,
        'discussion_identique': a['discussionIdentique'] if '@xsi:nil' not in a['discussionIdentique'] else np.nan,
    }

amendements = (
    pd.DataFrame.from_dict([ parse_amendement(a) for a in amendements_json ])
)

amendements

amendements.query('~seance.isna()')

amendements.duplicated(subset='numero').sum()

amendements.query('~discussion_identique.isna()')

amendements.query('~discussion_commune.isna()')

amendements.to_csv('out/amendements.csv', index=False)

dagstermill.yield_result(amendements, output_name="amendements")
