==> an-l16/1_prep_interventions.ipynb <==
import xml.dom.minidom
from glob import glob
import html
import unicodedata

import pandas as pd

import dagstermill

debats_xml =  glob('data/debats/xml/compteRendu/*.xml')

len(debats_xml)

reunions = pd.read_csv('out/reunions.csv')

def getText(nodelist):
    # Iterate all Nodes aggregate TEXT_NODE
    rc = []
    for node in nodelist:
        if node.nodeType == node.TEXT_NODE:
            rc.append(node.data)
        else:
            # Recursive
            rc.append(getText(node.childNodes))
    return ''.join(rc)

def get_interventions(debat_xml):
    doc = xml.dom.minidom.parse(debat_xml)
    paragraphes = doc.getElementsByTagName("paragraphe")
    seance = doc.getElementsByTagName('seanceRef')[0].childNodes[0].nodeValue
    interventions = []
    
    for p in paragraphes:
        id_syceron = p.getAttribute('id_syceron')
        acteur = p.getAttribute('id_acteur')
        code_grammaire = p.getAttribute('code_grammaire')
        code_style = p.getAttribute('code_style')
        #texte = p.getElementsByTagName('texte')[0]
        stime = p.getElementsByTagName('texte')[0].getAttribute('stime')
        #texte =  unicodedata.normalize("NFKD", " ".join([ n.nodeValue for n in p.getElementsByTagName('texte')[0].childNodes if n.nodeValue != None ]))
        texte =  unicodedata.normalize("NFKD", getText(p.getElementsByTagName('texte')))

        interventions.append({
            'id_syceron': id_syceron,
            'acteur': acteur,
            'code_grammaire': code_grammaire,
            'code_style': code_style,
            'stime': stime,
            'texte': texte
        })
    
    return (
        pd
        .DataFrame.from_records(interventions)
        .query('acteur != ""')
        .assign(seance = seance)
    )
    #return len(paragraphes)

interventions = pd.concat([get_interventions(debat) for debat in debats_xml])

interventions

interventions.query('id_syceron == "2913033"')

interventions_extra = (
    interventions
    .assign(
        mots = lambda df: df.texte.apply(lambda texte: len(texte.split(' ')))
    )
)

interventions_extra

# 9046284
# 9249922
interventions_extra.mots.sum()

interventions_extra.to_csv('out/interventions.csv', index=False)

dagstermill.yield_result(interventions_extra, output_name="interventions")

reunions = (
    reunions
    .set_index('uid')
)

(
    interventions_extra
    .join(reunions[['cr']], on='seance')
    .assign(
        url_seance = lambda df: df.cr.apply(lambda id: f'https://www.assemblee-nationale.fr/dyn/16/comptes-rendus/seance/{id}'),
    )
    .assign(
        url_intervention = lambda df: df.url_seance+'#'+df.id_syceron.astype(str),
    )
)
