==> an-l16/1_prep_references.ipynb <==
import json
from glob import glob
import pandas as pd

import dagstermill

def get_acteur(fn):
    data = json.load(open(fn))
    return {
        'uid': data['acteur']['uid']['#text'],
        'nom': data['acteur']['etatCivil']['ident']['nom'],
        'prenom': data['acteur']['etatCivil']['ident']['prenom'],
        'civ': data['acteur']['etatCivil']['ident']['civ'],
    }

acteurs = pd.DataFrame.from_records([ get_acteur(fn) for fn in glob('data/json/**/PA*.json') ])

acteurs

dagstermill.yield_result(acteurs, output_name="acteurs")

acteurs.to_csv('out/acteurs.csv', index=False)

def get_organe(fn):
    data = json.load(open(fn))['organe']
    return data

organes = pd.DataFrame.from_records([ get_organe(fn) for fn in glob('data/json/**/PO*.json') ])

organes

organes.to_csv('out/organes.csv', index=False)

dagstermill.yield_result(organes, output_name="organes")
