==> an-l16/1_prep_reunions.ipynb <==
import pandas as pd
import numpy as np
import json
import pandas as pd
from glob import glob
import tqdm

import dagstermill

def get_reunions(filename):
    return json.load(open(filename))['reunion']

reunions_json = [ get_reunions(f) for f in glob('data/reunions/json/**/RU*.json') ]

len(reunions_json)

reunions_ = (
    pd.DataFrame.from_dict(pd.json_normalize(reunions_json))
)

reunions_.columns

reunions_.query('~`participants.participantsInternes.participantInterne`.isna()')#[['participants.participantsInternes.participantInterne']]

def parse_reunion(r):
    return {
        'id': a['uid'],
        'numero': a['identification']['numeroLong'],
        'texte': a['texteLegislatifRef'],
        'signataire': a['signataires']['auteur']['acteurRef'],
        'organe': a['signataires']['auteur']['groupePolitiqueRef'],
        'dispositif' : a['corps']['contenuAuteur']['dispositif'] if 'contenuAuteur' in a['corps'] and 'dispositif' in a['corps']['contenuAuteur'] else np.nan,
        'cr': a['compteRenduRef'],
    }

reunions = (
    pd
    .DataFrame(
        #index = reunions_.uid
    )
    .assign(
        uid = reunions_.uid,
        nom=reunions_['demandeur.acteur.nom'],
        demandeur = reunions_['demandeur.acteur.acteurRef'],
        organe = reunions_['demandeur.organe.organeRef'],
        organeReuniRef = reunions_.organeReuniRef,
        participants = reunions_['participants.participantsInternes.participantInterne'],
        cr = reunions_['compteRenduRef']
    )
)

reunions

reunions.to_csv('out/reunions.csv', index=False)

dagstermill.yield_result(reunions, output_name="reunions")

reunions.organe.value_counts()

reunions.organeReuniRef.value_counts()

reunions.demandeur.value_counts()
