==> an-l16/1_prep_scrutins_votes.ipynb <==
positions = [
    'pours',
    'contres',
    'abstentions'
]

import json
import pandas as pd
from glob import glob
import tqdm

import dagstermill

def get_scrutin(filename):
    return json.load(open(filename))

scrutins = [ get_scrutin(f) for f in tqdm.tqdm(glob('data/**/VTAN*.json')) ]

len(scrutins)

def get_votes(scrutin):
    data = scrutin
    groupes = data['scrutin']['ventilationVotes']['organe']['groupes']['groupe']

    #print(scrutin)
    
    votes = [ (p, g['organeRef'], g['vote']['decompteNominatif'][p]['votant'], g['vote']['positionMajoritaire'])
           for g in groupes
           for p in positions
           if g['vote']['decompteNominatif'][p] != None
          ]

    df = (
        pd.concat([
            pd
            .DataFrame
            .from_records([v] if not isinstance(v, list) else v)
            .assign(
                position=p,
                organe=o,
                date=scrutin['scrutin']['dateScrutin'],
                scrutin=scrutin['scrutin']['uid'],
                groupe_majorite= gv
            )
            for (p, o, v, gv) in votes
        ])
        .assign(
            position = lambda df: df.position.replace({'pours': 'pour', 'contres': 'contre', 'abstentions': 'abstention'})
        )
    )

    return df

votes = pd.concat([ get_votes(scrutin) for scrutin in scrutins ])

votes

votes.to_csv('out/votes.csv', index=False)

votes.to_parquet('out/votes.par', index=False)

dagstermill.yield_result(votes, output_name="votes")

#scrutins[0]

scrutins_df = pd.DataFrame.from_records([
    {
        'uid': s['scrutin']['uid'],
        'objet': s['scrutin']['titre'],
        'type': s['scrutin']['typeVote']['libelleTypeVote'],
        'date': s['scrutin']['dateScrutin'],
        'seance': s['scrutin']['seanceRef'],
        'sort': s['scrutin']['sort']['code']
    }
    for s in scrutins
])

scrutins_df

scrutins_df.type.value_counts()

scrutins_df.to_csv('out/scrutins.csv', index=False)

dagstermill.yield_result(scrutins_df, output_name="scrutins")
